Add tests
This version of the test script uses the test command to check if the output of the script is as expected. You can run this test script in the same way as before by saving it in a file with a .sh extension, for example test_script.sh and then running the command bash test_script.sh on the terminal.

Also updated README